INKLB-014 separate convenience header
